+++ crates/affected/src/affected_tracker.rs
@@ -148,12 +148,6 @@ impl AffectedTracker {
     }
 
     pub fn is_project_affected(&self, project: &Project) -> Option<AffectedBy> {
-        // If the project was marked with a legitimate source that isn't relationship
-        // related, then we can short-circuit early
-        if self.is_project_marked_ignoring_relations(project) {
-            return Some(AffectedBy::AlreadyMarked);
-        }
-
         if project.is_root_level() {
             // If at the root, any file affects it
             self.changed_files
@@ -353,12 +347,6 @@ impl AffectedTracker {
     }
 
     pub fn is_task_affected(&self, task: &Task) -> miette::Result<Option<AffectedBy>> {
-        // If the task was marked with a legitimate source that isn't relationship
-        // related, then we can short-circuit early
-        if self.is_task_marked_ignoring_relations(task) {
-            return Ok(Some(AffectedBy::AlreadyMarked));
-        }
-
         // Special CI handling
         match (self.ci, &task.options.run_in_ci) {
             (true, TaskOptionRunInCI::Always) => {